S15.f.2: extract plugin.ts delegateToNativeAgent (6974 → 6890 LOC)
Second sub-slice of the S15.f plugin.ts decomposition train. Lifts
the AGT task-request → native OpenClaw agent dispatcher into a
dedicated module.

cli/src/core/agt-task-delegate.ts (95 LOC):
  delegateToNativeAgent(taskContent, fromAgent, log) — pure stdlib
  helper. Zero plugin-internal dependencies (no AGT/router/cache
  singletons). The 'log' param is already injected by callers, so
  this is the cleanest extraction in the train.

Behavioral change: NONE. Body byte-identical.

Cumulative S15.f delta:
  pre:  7139
  f.1:  6974  (-165)
  f.2:  6890   (-84)
  Cumulative: -249. 6090 LOC remaining to §4.2 cap.

Verification:
  - tsc --noEmit clean / lint 24 (unchanged) / build clean
  - npm test: 454 pass / 2 skipped pre-existing (same as baseline)

Audit: docs/security-audits/2026-04-29-phase2-hotspot-plugin-cli-f2.md
